Update intro component to accept descriptionContent as a property
